1) Messages controller has been improved 2) Messages example has been refactored and fixed 3) Implementations were  moved to cpp

+++ include/controller/base.h
@@ -25,7 +25,7 @@ namespace Textmagic {
 		public:
             typedef struct
                 {
-                  int code;
+                  long code;
                   std::string message;
                   Json::Value errors;
                 } Error;
